// web/src/api/generated/model/resultsState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { SittingOut } from './sittingOut';
import type { ZimsecSubjectOut } from './zimsecSubjectOut';

export interface ResultsState {
  sittings: SittingOut[];
  subjects: ZimsecSubjectOut[];
  name_on_id: string | null;
}